feat: Fix bug in game.py timer
The code changes in game.py fix a bug related to the timer functionality. Previously, the timer would continue even if the game instance was no longer active. With this fix, the timer will only send a message and skip the next player if the game instance is still active.

+++ unu/utils.py
@@ -37,5 +37,6 @@ async def load_all():
         await game.delete()
 
     for player in await GamePlayer.all():
-        player_game[player.player_id] = games[player.game_chat_id]
-        await player.delete()
+        if player.game_chat_id in games:
+            player_game[player.player_id] = games[player.game_chat_id]
+            await player.delete()
